+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v3.0.3
 * Added: Added a hasAuthorization method to the Query class to allow searches for any element that uses an authorization string or strings.
 * Fixed: GeoPoint.distanceBetween and GeoPoint.calculateCenter calculations
+* Added: Query extended data on an element
 
 # v3.0.2
 * Fixed: Elasticsearch5 throwing unsupported operation exception when adding/updating a property